[JAVA-47895] Moved code from core-java-networking-5 to core-java-networking-6
