+++ core-java-modules/core-java-networking-6/src/test/java/com/baeldung/getwebfilesize/GetWebFileSizeLiveTest.java
@@ -1,11 +1,13 @@
 package com.baeldung.getwebfilesize;
 
-import org.junit.jupiter.api.Test;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.fail;
+
 import java.io.IOException;
 import java.net.URL;
 import java.net.URLConnection;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.fail;
+
+import org.junit.jupiter.api.Test;
 
 class GetWebFileSizeLiveTest {
